+++ oxe-api/resource/campaign/add_campaign_template.py
@@ -1,0 +1,29 @@
+from flask_apispec import MethodResource
+from flask_apispec import doc
+from flask_jwt_extended import jwt_required
+from flask_restful import Resource
+
+from decorator.catch_exception import catch_exception
+from decorator.log_request import log_request
+from decorator.verify_admin_access import verify_admin_access
+
+
+class AddCampaignTemplate(MethodResource, Resource):
+
+    def __init__(self, db):
+        self.db = db
+
+    @log_request
+    @doc(tags=['campaign'],
+         description='Add a campaign template',
+         responses={
+             "200": {},
+         })
+    @jwt_required
+    @verify_admin_access
+    @catch_exception
+    def post(self):
+
+        template = self.db.insert({}, self.db.tables["CampaignTemplate"])
+
+        return template.id, "200 "