+++ oxe-web-admin/src/component/Menu.jsx
@@ -108,7 +108,7 @@ export default class Menu extends React.Component {
 							<i className="fas fa-shapes" style={{ fontSize: "1.75em" }} />
 						</NavIcon>
 						<NavText>
-							Entity
+							Entities
 						</NavText>
 					</NavItem>
 					<NavItem
@@ -147,16 +147,16 @@ export default class Menu extends React.Component {
 						</NavItem>
 					}
 					<div className="Menu-divider"/>
-					{getSettingValue(this.props.settings, "SHOW_COMMUNICATION_PAGE") === "TRUE"
+					{getSettingValue(this.props.settings, "SHOW_CAMPAIGN_PAGE") === "TRUE"
 						&& <NavItem
-							eventKey="communication"
-							active={this.props.selectedMenu === "communication"}
-							onClick={() => this.props.history.push("/communication")}>
+							eventKey="campaign"
+							active={this.props.selectedMenu === "campaign"}
+							onClick={() => this.props.history.push("/campaign")}>
 							<NavIcon>
-								<i className="fas fa-bullhorn" style={{ fontSize: "1.75em" }} />
+								<i className="fas fa-mail-bulk" style={{ fontSize: "1.75em" }} />
 							</NavIcon>
 							<NavText>
-								Communication via email
+								Email Campaigns
 							</NavText>
 						</NavItem>
 					}
@@ -169,7 +169,7 @@ export default class Menu extends React.Component {
 								<i className="fas fa-poll-h" style={{ fontSize: "1.75em" }} />
 							</NavIcon>
 							<NavText>
-								Form
+								Forms
 							</NavText>
 							{this.getFormNotificationBlock()}
 						</NavItem>
@@ -182,7 +182,7 @@ export default class Menu extends React.Component {
 							<i className="fas fa-photo-video" style={{ fontSize: "1.75em" }} />
 						</NavIcon>
 						<NavText>
-							Media
+							Images & Documents
 						</NavText>
 					</NavItem>
 					<div className="Menu-divider"/>