+++ oxe-web-community/src/component/Login.jsx
@@ -129,7 +129,7 @@ export default class Login extends React.Component {
 		};
 
 		postRequest.call(this, "account/forgot_password", params, () => {
-			nm.info("An email has been sent with a link to reset your password");
+			nm.info("If that email address is in our database, we will send you an email to reset your password");
 		}, (response) => {
 			nm.warning(response.statusText);
 		}, (error) => {
@@ -149,6 +149,11 @@ export default class Login extends React.Component {
 		}, (error) => {
 			nm.error(error.message);
 		});
+	}
+
+	backToLogin() {
+		this.props.cookies.remove("access_token_cookie");
+		window.location.replace("/");
 	}
 
 	onKeyDown(event: React.KeyboardEvent<HTMLDivElement>) {
@@ -514,7 +519,7 @@ export default class Login extends React.Component {
 									<div className="left-buttons">
 										<button
 											className="link-button"
-											onClick={() => this.changeState("view", "login")}
+											onClick={() => this.backToLogin()}
 										>
 											Back to login
 										</button>
@@ -549,7 +554,7 @@ export default class Login extends React.Component {
 								<div className="left-buttons">
 									<button
 										className="link-button"
-										onClick={() => this.changeState("view", "login")}
+										onClick={() => this.backToLogin()}
 									>
 										Back to login
 									</button>
@@ -609,7 +614,7 @@ export default class Login extends React.Component {
 								<div className="left-buttons">
 									<button
 										className="link-button"
-										onClick={() => window.location.replace("/")}
+										onClick={() => this.backToLogin()}
 									>
 										Back to login
 									</button>